requirements.txt
fastapi
uvicorn
python-dotenv
stripe
transformers
torch
accelerate
llama-cpp-python
